requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-filter==23.5
django-import-export==3.3.4
django-jazzmin==2.6.0
django-unfold==0.18.0
dotty-dict==1.3.1
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dotenv==1.0.0
python-gitlab==4.2.0
python-semantic-release==8.5.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
shellingham==1.5.4
smmap==5.0.1
sqlparse==0.4.4
tablib==3.5.0
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
